requirements.txt
numpy==1.20.1
pycocotools==2.0.2
matplotlib==3.3.4
torchvision==0.2.2
scipy==1.6.1
tqdm==4.19.9
torch==0.4.1.post2
Pillow==8.1.0
tensorboardX==2.1
tensorboard==2.2.2
